tms_backend/requirements.txt
Django==5.2.7
psycopg2-binary==2.9.10
gunicorn==23.0.0
django-environ==0.11.2
djangorestframework==3.15.2
django-cors-headers==4.4.0